Rename component js files to index js files, update relevant paths
