--- src/components/Button/Button.css
@@ -1,69 +1,0 @@
-.Button {
-  font-family: 'News Cycle', sans-serif;
-  height: 45px;
-  width: 150px;
-  border: 2px solid gold;
-  background-color: transparent;
-  color: gold;
-  font-size: 1.25rem;
-  border-radius: 15px;
-  padding-bottom: 3px;
-  letter-spacing: 1px;
-  text-decoration: none;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-}
-
-.Button:hover {
-  color: white;
-  border-color: white;
-}
-
-.favorites {
-  height: 75px;
-  max-width: 100px;
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  padding-bottom: 7px;
-}
-
-.favorites:hover {
-  box-shadow: 0 0 15px gold, 0 0 15px white;
-}
-
-.favorites-badge {
-  font-size: 1.75rem;
-  font-weight: 700;
-  margin: 0;
-  line-height: 1;
-}
-
-.people:hover {
-  box-shadow: 0 0 15px deepskyblue, 0 0 15px white;
-}
-
-.planets:hover {
-  box-shadow: 0 0 15px springgreen, 0 0 15px white;
-}
-
-.vehicles:hover {
-  box-shadow: 0 0 15px red, 0 0 15px white;
-}
-
-.active {
-  background-color: gold;
-  color: black;
-  font-weight: 700;
-}
-
-.active:hover {
-  color: black;
-  border-color: black;
-}
-
-.active:focus {
-  outline: none;
-}
