+++ src/components/ScrollingText/index.js
